# src/main/resources/openapi/restaurants-api.yaml
openapi: 3.0.1
info:
  title: Restaurants Microservice API
  description: API for restaurant management in Plaza de Comidas system
  version: 1.0.0
  contact:
    name: Mateo Castro
    email: mateocastro1297@gmail.com

servers:
  - url: http://localhost:8082/api/v1/restaurants
    description: Local server

tags:
  - name: Restaurants
    description: Restaurant management operations
  - name: Dishes
    description: Dish management operations for restaurants

paths:
  /admin/restaurant:
    post:
      summary: Create a new restaurant
      operationId: createRestaurant
      tags:
        - Restaurants
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-restaurant.yaml#/components/schemas/RestaurantRequest'
      responses:
        '200':
          description: Restaurant created successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-restaurant.yaml#/components/schemas/RestaurantResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - NIT already exists
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /owner/dish:
    post:
      summary: Create a new dish
      operationId: createDish
      tags:
        - Dishes
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-dish.yaml#/components/schemas/DishRequest'
      responses:
        '200':
          description: Dish created successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-dish.yaml#/components/schemas/DishResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not the restaurant owner
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Restaurant or category not found
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /owner/dish/{dishId}:
    put:
      summary: Update dish price and description
      operationId: updateDish
      tags:
        - Dishes
      security:
        - bearerAuth: [ ]
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-dish.yaml#/components/schemas/DishUpdateRequest'
      responses:
        '200':
          description: Dish updated successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-dish.yaml#/components/schemas/DishResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not the restaurant owner
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Dish not found
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /owner/dish/{dishId}/toggle-active:
    patch:
      summary: Enable or disable a dish
      operationId: toggleDishActive
      tags:
        - Dishes
      security:
        - bearerAuth: [ ]
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dish status toggled successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-dish.yaml#/components/schemas/DishResponse'
        '403':
          description: Forbidden - User is not the restaurant owner
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '404':
          description: Dish not found
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /owner/employee:
    post:
      summary: Associate employee to restaurant
      operationId: createRestaurantEmployee
      tags:
        - Restaurants
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-restaurant-employee.yaml#/components/schemas/RestaurantEmployeeRequest'
      responses:
        '200':
          description: Employee associated successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-restaurant-employee.yaml#/components/schemas/RestaurantEmployeeResponse'
        '403':
          description: Forbidden - User is not the restaurant owner
        '404':
          description: Restaurant not found
        '409':
          description: Employee already associated to this restaurant

  /clients/restaurants:
    get:
      summary: List all restaurants paginated and alphabetically ordered
      operationId: listRestaurants
      tags:
        - Restaurants
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Restaurants listed successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-restaurant.yaml#/components/schemas/PagedRestaurantResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /clients/restaurants/{restaurantId}/dishes:
    get:
      summary: List dishes of a restaurant filtered by category
      operationId: listDishesByRestaurant
      tags:
        - Dishes
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Dishes listed successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-dish.yaml#/components/schemas/PagedDishResponse'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Enter your JWT token in the format: Bearer <token>"